split code into smaller components
