+++ JS/js-structures/js-structure_shapes/index.js
@@ -1,23 +1,13 @@
 console.clear();
 
+import Circle from "./components/Circle/Circle.js";
+import Pentagon from "./components/Pentagon/Pentagon.js";
+import Square from "./components/Square/Square.js";
+
 const root = document.getElementById("root");
 
-const circle = document.createElement("div");
-circle.classList.add("circle");
-circle.addEventListener("click", () => {
-  circle.style.backgroundColor = "#ccc";
-});
-
-const square = document.createElement("div");
-square.classList.add("square");
-square.addEventListener("click", () => {
-  square.style.backgroundColor = "#ccc";
-});
-
-const pentagon = document.createElement("div");
-pentagon.classList.add("pentagon");
-pentagon.addEventListener("click", () => {
-  pentagon.style.backgroundColor = "#ccc";
-});
+const circle = Circle();
+const pentagon = Pentagon();
+const square = Square();
 
 root.append(circle, square, pentagon);
